[Refactor(christmas)] 패키지 이동
